temp/ccu/ccu/ccu_grpc/requirements.txt
grpcio>=1.53.0
protobuf>=4.22.1
